Backstab Mechanic 99% Complete
When attacking from behind with a dagger, you will "backstab" the target entity. Marked as 99% complete because it needs a tiny bit of cleanup to stop printing out debug text and fix some border cases in which when both entities in the interaction are facing the direction '0' the backstab mechanic will go off even if the entities do not fit the original criteria for a backstab to happen. (Enemies that attack eachother with a dagger within 30degrees yaw of one another execute the backstab).